perf(decode): R12 monolithic per-kernel decoder foundation (#291)
* refactor(decode): AVX2-tier monolithic sequence decoder

Replaces the macro-generated three-function tier (decode_fn / loop_fn /
decode_one_fn) for the AVX2 path with one self-contained
`#[target_feature(enable = "bmi2,avx2")]` function. The body inlines the
entire decode + execute pipeline: outer init, RLE dispatch, FSE state
init, both pipeline arms (8-deep lookahead ring for cold-dict / >16 MB
total-history, straight short-block loop otherwise), per-sequence
decode (peek_bits_triple_bmi2 with `_pext_u64` inline at call site),
per-sequence execute (32-byte AVX2 ymm wildcopy via
exec_sequence_inline_avx2), and post-loop bitstream-tail validation.

Two internal helpers (decode_one_avx2, execute_one_avx2) carry the same
target_feature scope and are `#[inline]` — LLVM collapses them into the
caller in release mode since both are single-callsite hot-path functions
inside the matching target_feature.

Companion monoliths for BMI2 / VBMI2 / Scalar in follow-up commits. The
K-generic dispatcher in sequence_section_decoder still selects the tier
once per call via cached detect_cpu_kernel.

* refactor(decode): BMI2-tier monolithic sequence decoder

Same shape as the AVX2 monolith but pinned to `Bmi2Kernel` and using
the SSE2 16-byte `exec_sequence_inline` for match copy (no AVX2 ymm
widening at this tier). Outer init + RLE dispatch + FSE state init +
both pipeline arms + decode_one + execute_one all inlined into one
`#[target_feature(enable = "bmi2")]` function.

Triple-bit extract goes through `peek_bits_triple_bmi2` (_pext_u64
inline) gated by the cached vendor policy — same as AVX2.

* refactor(decode): VBMI2-tier monolithic sequence decoder

Same shape as the AVX2 monolith but pinned to `Vbmi2Kernel` and
carrying the full VBMI2 + AVX-512 + AVX2 + BMI2 target_feature scope.
Match copy uses `exec_sequence_inline_avx2` since VBMI2 always implies
AVX2+BMI2; zmm 64-byte wildcopy + VPSHUFB short-offset shortcut wait
until WILDCOPY_OVERLENGTH grows to 64.

* refactor(decode): Scalar monolith + drop x86 macro + one-time dispatch

Adds `seq_decoder_scalar.rs` as the Scalar-tier monolith (same shape
as the AVX2 / BMI2 / VBMI2 monoliths, pinned to `ScalarKernel`, no
target_feature). Dispatcher in
`sequence_section_decoder::decode_and_execute_sequences` now routes
the Scalar arm into the new monolith too.

Deletes `seq_decoder_x86_kernel.rs` entirely — the
`define_x86_seq_decoder_tier!` macro is no longer referenced by any
tier file. Each tier (Scalar / BMI2 / AVX2 / VBMI2) is now a fully
self-contained function with the whole decode + execute pipeline
inlined in one place, selected ONCE per call via cached
`detect_cpu_kernel`.

aarch64 NEON / SVE still go through the K-generic
`decode_and_execute_sequences_impl` shared body until their own
monoliths land in a follow-up.

* perf(decode): macro-expand decode_one + execute_one inside each kernel monolith

Replaces the per-tier `#[inline]` helper functions (`decode_one_avx2`,
`execute_one_avx2`, BMI2/VBMI2/Scalar equivalents) with `macro_rules!`
blocks that expand textually at every callsite inside the monolithic
outer function. This bypasses the LLVM inline cost-model entirely:
macros expand BEFORE LLVM sees the code, so the resulting compiled
function has NO inner CALL boundary regardless of cost-model decisions.

Why this matters: `#[inline]` is a hint, not a command. With
`target_feature` scopes, `Result<>` panic landings, multiple callsites,
and generic `B: BufferBackend` parameterisation, LLVM's heuristic
declines to inline — leaving CALL boundaries that fragment register
allocation and break the µop-cache window on Skylake-X
(DSB ≤ 1536 µops). Combined with the Rust limit that forbids
`#[inline(always)]` + `#[target_feature]` together (rust-lang/rust#145574),
macros are the only mechanism that guarantees the textual inline donor C
achieves via single-source monoliths.

Macro shape: two `macro_rules!` per tier — `decode_one_body!`
(state read + triple-bit extract + advance) and `execute_one_body!`
(literal slice + offset gates + match copy + cold fallback). The execute
body uses labeled-block early-exit (`break 'exec_inner Err(...)`),
stable since 1.65, instead of closure or `?` — closures and `?` recreate
function boundaries.

Affects all four monoliths: Scalar / BMI2 / AVX2 / VBMI2.

* perf(huff0): drop Vec<Entry> table, unify on packed u16 (donor HUF_DEltX1) (#292)

The HUF decode table was held in two parallel representations:
- `decode: Vec<Entry { symbol: u8, num_bits: u8 }>` for scalar /
  weight-stream / single-symbol fallback readers
- `packed_decode: Vec<u16>` (donor `HUF_DEltX1` format: low byte =
  symbol, high byte = num_bits) for the hot 4-stream burst path

`build_table_from_weights` filled both with `Vec::fill(value)` over
the same index range — two `__memset_avx2`-driven passes per rank
group. On the z000033 decode flamegraph the HUF table build is
12.53% of self-time, with `build_table_from_weights` 5.02% (largest
single setup cost on a workload with many small blocks).

Drop the `Entry` Vec entirely. Every reader switches to inline
unpack of the packed u16: `packed as u8` for symbol, `(packed >> 8)
as u8` for num_bits. Single AND/SHR on registers — same cost as the
struct-field reads being replaced. Donor `huf_decompress.c` reads
the `HUF_DEltX1` byte-wise from a single u16-wide table, so this is
structural parity rather than a representation invention.

`Entry` struct dropped as unused; no external re-export.
