+++ zstd/src/decoding/mod.rs
@@ -55,22 +55,20 @@ pub(crate) mod prefetch;
 mod ringbuffer;
 #[allow(dead_code)]
 pub(crate) mod scratch;
-// Per-kernel sequence-section decoder entry points. Each module owns
-// its full `#[target_feature]`-tagged decoder body generated by the
-// `define_x86_seq_decoder_tier!` macro in `seq_decoder_x86_kernel`.
-// Per-tier bodies carry the BMI2-direct `peek_bits_triple_bmi2` hot
-// site (#279 round 3 Phase 3); the AVX2 tier additionally routes
-// match-copy through `wildcopy_no_overlap_avx2` (32-byte ymm stride)
-// via `BufferBackend::exec_sequence_inline_avx2` (#279 round 3
-// Phase 4). The shared K-generic `decode_and_execute_sequences_impl`
-// remains for the Scalar arm and aarch64 dispatch only.
-#[cfg(target_arch = "x86_64")]
-#[macro_use]
-pub(crate) mod seq_decoder_x86_kernel;
+// Per-kernel monolithic sequence-section decoder entry points. Each
+// kernel has its own self-contained function with the full pipeline
+// (outer init, both arms, decode_one, execute_one) inlined inside one
+// `#[target_feature]`-scoped body. The dispatcher in
+// `sequence_section_decoder::decode_and_execute_sequences` selects the
+// kernel ONCE per call via cached `detect_cpu_kernel`. aarch64 Neon
+// and Sve still go through the K-generic
+// `decode_and_execute_sequences_impl` shared body until their own
+// monoliths land.
 #[cfg(target_arch = "x86_64")]
 pub(crate) mod seq_decoder_avx2;
 #[cfg(target_arch = "x86_64")]
 pub(crate) mod seq_decoder_bmi2;
+pub(crate) mod seq_decoder_scalar;
 #[cfg(target_arch = "x86_64")]
 pub(crate) mod seq_decoder_vbmi2;
 pub(crate) mod sequence_execution;
